DC-144: Check health endpoint during cert import process: fix PR comments
