+++ plugin/src/test/java/org/opennms/plugins/cloud/tsaas/config/ConfigZipExtractorTest.java
@@ -26,30 +26,24 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.plugins.cloud.tsaas.shell;
+package org.opennms.plugins.cloud.tsaas.config;
 
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Set;
+import static org.junit.Assert.*;
 
-import org.opennms.integration.api.v1.scv.Credentials;
-import org.opennms.integration.api.v1.scv.SecureCredentialsVault;
+import java.io.IOException;
+import java.nio.file.Path;
 
-public final class InMemoryScv implements SecureCredentialsVault {
-    final Map<String, Credentials> creds = new HashMap<>();
+import org.junit.Test;
 
-    @Override
-    public Set<String> getAliases() {
-        return creds.keySet();
+public class ConfigZipExtractorTest {
+
+    @Test
+    public void shouldExtract() throws IOException {
+        Path configZipFile = Path.of("src/test/resources/cert/cloud-credentials.zip");
+        ConfigZipExtractor ex = new ConfigZipExtractor(configZipFile);
+        assertTrue(ex.getPrivateKey().startsWith("-----BEGIN PRIVATE KEY-----"));
+        assertTrue(ex.getPublicKey().startsWith("-----BEGIN CERTIFICATE-----"));
+        assertTrue(ex.getJwtToken().startsWith("acme"));
     }
 
-    @Override
-    public Credentials getCredentials(String s) {
-        return creds.get(s);
-    }
-
-    @Override
-    public void setCredentials(String s, Credentials credentials) {
-        creds.put(s, credentials);
-    }
 }
